Make FEServer not extend Game
Most of Game's functionality is tailored towards managing a LWJGL
window. FEServer doesn't have a LWJGL window. In practice, this meant
that the only part of Game that FEServer used was the timeDelta
variable, and even that was used in a write-only fashion. And FEServer's
implementation of `loop` was drastically different than the `loop`
implementations of the other two classes that extend Game: FEMultiplayer
and LevelEditor. Basically, this inheritance doesn't really fit.